requirements.txt
streamlit
streamlit_option_menu
gtts
PyPDF2
google-generativeai
python-dotenv==1.0.1
langchain==0.2.10
langchain-core
langchain-community
langchain-google-genai
streamlit-option-menu
joblib
groq
qrcode==7.4.2
faiss-cpu==1.7.2
unstructured==0.15.0
unstructured[pdf]==0.15.0
transformers==4.43.3
sentence-transformers==3.0.1
nltk==3.8.1
mtranslate
pycryptodome
bcrypt